fixed table import
