+++ common/pom.xml
@@ -52,10 +52,10 @@
       <groupId>com.auth0</groupId>
       <artifactId>java-jwt</artifactId>
     </dependency>
-      <dependency>
-          <groupId>org.jooq</groupId>
-          <artifactId>jooq</artifactId>
-      </dependency>
+    <dependency>
+      <groupId>org.jooq</groupId>
+      <artifactId>jooq</artifactId>
+    </dependency>
   </dependencies>
 
 </project>
