+++ service/src/main/java/com/codeforcommunity/processor/SiteProcessorImpl.java
@@ -12,11 +12,12 @@ import com.codeforcommunity.dto.site.GetSiteResponse;
 import com.codeforcommunity.dto.site.SiteEntry;
 import com.codeforcommunity.dto.site.StewardshipActivitiesResponse;
 import com.codeforcommunity.dto.site.StewardshipActivity;
-import com.codeforcommunity.benefits.TreeBenefitsCalculator;
 import com.codeforcommunity.dto.site.TreeBenefitsResponse;
 import com.codeforcommunity.enums.SiteOwner;
 import com.codeforcommunity.exceptions.ResourceDoesNotExistException;
 import com.codeforcommunity.logger.SLogger;
+import com.codeforcommunity.requester.TreeBenefitsCalculator;
+
 import java.util.ArrayList;
 import java.util.List;
 import org.jooq.DSLContext;
@@ -237,30 +238,28 @@ public class SiteProcessorImpl implements ISiteProcessor {
   public TreeBenefitsResponse calculateBenefits(int siteId) {
     checkSiteExists(siteId);
 
-    SiteEntriesRecord record = db.selectFrom(SITE_ENTRIES)
-            .where(SITE_ENTRIES.SITE_ID.eq(siteId))
-            .fetchOne();
+    SiteEntriesRecord record =
+        db.selectFrom(SITE_ENTRIES).where(SITE_ENTRIES.SITE_ID.eq(siteId)).fetchOne();
 
-    if (record==null) {
+    if (record == null) {
       throw new ResourceDoesNotExistException(siteId, "site entry");
     }
 
     String commonName = record.getCommonName();
     double diameter = record.getDiameter();
 
-    if (commonName==null) {
+    if (commonName == null) {
       throw new ResourceDoesNotExistException(siteId, "site entry common name");
-    } else if (diameter==0) {
+    } else if (diameter == 0) {
       throw new ResourceDoesNotExistException(siteId, "site entry diameter");
     }
 
     TreeBenefitsCalculator calculator = new TreeBenefitsCalculator(this.db, commonName, diameter);
     return new TreeBenefitsResponse(
-            calculator.calcEnergy(), calculator.calcEnergyMoney(),
-            calculator.calcStormwater(), calculator.calcStormwaterMoney(),
-            calculator.calcAirQuality(), calculator.calcAirQualityMoney(),
-            calculator.calcCo2Removed(), calculator.calcCo2RemovedMoney(),
-            calculator.calcCo2Stored(), calculator.calcCo2StoredMoney()
-    );
+        calculator.calcEnergy(), calculator.calcEnergyMoney(),
+        calculator.calcStormwater(), calculator.calcStormwaterMoney(),
+        calculator.calcAirQuality(), calculator.calcAirQualityMoney(),
+        calculator.calcCo2Removed(), calculator.calcCo2RemovedMoney(),
+        calculator.calcCo2Stored(), calculator.calcCo2StoredMoney());
   }
 }
